Extracting StoreInDevAddrCache in different class
